# Writer: switch back to SemaphoreSlim