@NickCraver 2022-01-18
Decision: we'll `#ifdef` around to only .NET Core 3.1+ this