@NickCraver 2022-01-18
...okay, I hate the perf loss, but this is also ridiculous.